requirements.txt
anyio==3.6.2
black==22.10.0
click==8.1.3
colorama==0.4.5
fastapi==0.85.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
pathspec==0.10.1
platformdirs==2.5.2
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0
